Rename Builder to AuthBuilder and create Builder wrapper
Previously, OmniAuth::Builder was a subclass of Rack::Builder.
Rack::Builder#call calls #to_app, rebuilding the Rack application on
each request, and, as it says in the Rack::Builder comments, this may
have a negative impact on performance.

This commit renames the original OmniAuth::Builder to OmniAuth::AuthBuilder
and replaces it with a wrapper to be used as middleware (as is suggested
in the documentation).

Using OmniAuth::Builder will now only build the application once.

Resolves #1083.
